+++ rust/operator-binary/src/main.rs
@@ -46,6 +46,7 @@ async fn main() -> anyhow::Result<()> {
             product_config,
             watch_namespace,
             tracing_target,
+            cluster_info_opts,
         }) => {
             stackable_operator::logging::initialize_logging(
                 "DRUID_OPERATOR_LOG",
@@ -64,8 +65,11 @@ async fn main() -> anyhow::Result<()> {
                 "deploy/config-spec/properties.yaml",
                 "/etc/stackable/druid-operator/config-spec/properties.yaml",
             ])?;
-            let client =
-                stackable_operator::client::create_client(Some(OPERATOR_NAME.to_string())).await?;
+            let client = stackable_operator::client::initialize_operator(
+                Some(OPERATOR_NAME.to_string()),
+                &cluster_info_opts,
+            )
+            .await?;
 
             Controller::new(
                 watch_namespace.get_api::<DeserializeGuard<DruidCluster>>(&client),